增加更新取消功能，修改笔迹步骤分模式切换图片显示问题，修改C5电子笔无法批注问题，增加回评列表时间和分数排序功能

+++ app/src/main/java/com/ruiyi/netreading/bean/SpenStroke.java
@@ -3,7 +3,7 @@ package com.ruiyi.netreading.bean;
 import android.graphics.PointF;
 
 public class SpenStroke {
-    private int penSize; //笔的宽度
+    private float penSize; //笔的宽度
     private String color; //笔的颜色
     private String pageName; //当前页面(A、B面)
     private PointF[] pointS; //触摸点
@@ -11,7 +11,7 @@ public class SpenStroke {
     //private int[] integerS; //时间戳
 
 
-    public int getPenSize() {
+    public float getPenSize() {
         return penSize;
     }
 
@@ -27,7 +27,7 @@ public class SpenStroke {
         this.pageName = pageName;
     }
 
-    public void setPenSize(int penSize) {
+    public void setPenSize(float penSize) {
         this.penSize = penSize;
     }
